Delivery as a service.
This commit changes the asset delivery to run as a service that can take
in any URL and convert the contents using query parameters. This also
does away with Amazon S3 support in favor of Google Storage. This has
been abstracted behind a FileSystem interface for ease of migration in
the future.

Additionally this can run as a Docker image, so have at it.