Asset Delivery & Resizing Separated.
Asset delivery was slow on first load of every image as it required waiting for the process to finish resizing before pointing the user to a URL from which the asset can be downloaded. This was fixed by returning the original URL if a resized version could not be found and performing resizing in a goroutine.

However, goroutines are automatically killed at the end of any request by Google Cloud Run (which is what we use to run the delivery). Therefore, this did not work.

This PR moves the resizing functionality to a public function which is managed by Google Cloud Functions. It is meant to be connected to a specific PubSub topic. The original delivery server now sends a message on that PubSub topic to start the process.

TODO: Externalize the topic in an environment variable

TODO: find a way to stop duplicate images from being created (perhaps we can use Redis/MemCached to create a key using the hash, the width, and the encoding format. This key could be deleted once encoding is finished, and the value of the key could be an expiry after which the function should attempt to recreate the file).