Fallback when deocding images
When reading images, we previously used the "hint" (e.g., the location URL) to try to determine the type of file we are expecting. However, sometimes, the extension of the file is incorrect. In those cases, asset delivery will constantly present with an error.

Now, if an error occurs during decoding, the system will attempt a more general (but slower) decoding process which has to go through all formats to see if it matches any of the formats.

https://app.asana.com/0/1202503093841789/1203168895393612/f